% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/beta.mom.R
\name{beta.mom}
\alias{beta.mom}
\title{Fits the beta distribution by method of moments.}
\usage{
beta.mom(dat)
}
\arguments{
\item{dat}{A vector contains the numbers that are assumed to follow a
beta distribution.}
}
\value{
Returns the estimation of alpha and beta.
}
\description{
Fits the beta distribution by method of moments.
}
\examples{
beta.mom(rbeta(5,5,20))
}
\author{
Ning Leng
}

