% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaempH0.R
\name{BetaempH0}
\alias{BetaempH0}
\title{Compare empirical non-directional KS statistics to KS statistics of permuted data}
\usage{
BetaempH0(Data, DataShuffle,StateNames=c("EE","DE"),PiIn=NULL,Num=-100, iter=5, UpdatePi=TRUE)
}
\arguments{
\item{Data}{directional KS statistics of empirical data}

\item{DataShuffle}{directional KS statistics of permuted data}

\item{StateNames}{names of the states}

\item{PiIn}{initial value for Pi}

\item{Num}{The underflowed values will be imputed as exp(Num)}

\item{iter}{number of iterations}

\item{UpdatePi}{whether update Pi}
}
\value{
Output: posterior probabilities and estimated hyper-parameters
}
\examples{
tmp1 <- runif(10)
mat1 <- cbind(tmp1, 1-tmp1)
tmp2 <- runif(10)
mat2 <- cbind(tmp2, 1-tmp2)
res <- BetaempH0(mat1, mat2)
}
\author{
Ning Leng
}
