% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DirichletOptim.R
\name{DirichletOptim}
\alias{DirichletOptim}
\title{Estimate parameters of a dirichlet distribution}
\usage{
DirichletOptim(ParamPool, InputPool)
}
\arguments{
\item{ParamPool}{parameters}

\item{InputPool}{input values}
}
\value{
sum likelihood
}
\description{
Estimate parameters of a dirichlet distribution
}
\author{
Ning Leng
}

