% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KSNoDir.R
\name{KSNoDir}
\alias{KSNoDir}
\title{Calculate non-directional KS-statistics of pair-wise comparisons}
\usage{
KSNoDir(SCMat, Cond, pval=FALSE,Dropout.remove=FALSE, Dropout.upper=0)
}
\arguments{
\item{SCMat}{A data matrix contains expression values for each transcript
(gene or isoform level). In which rows should be transcripts
and columns should be samples. The matrix is expected to be normalized
prior to the analysis.}

\item{Cond}{A factor indicates the condition which each sample belongs to.}

\item{pval}{whether output KS test raw p values}

\item{Droupout.remove,Dropout.upper}{For a given gene, whether cells with dropout will be
ignored in the test. Defaulte is FALSE. If it is set to TRUE, values that are less or equal to
Dropout.upper will be ignored during the test.}
}
\value{
Output contains KS statistics (non-directional)
for each pairwise comparison.
Note the order of conditions are defined by the order in the Cond factor.
}
\description{
Calculate non-directional KS-statistics of pair-wise comparisons
}
\examples{
mat <- matrix(rnorm(40), ncol=10)
tmp1 <- KSNoDir(mat, as.factor(rep(c("c1","c2"),each=5))
}
\author{
Ning Leng
}

