\name{SCPattern-package}
\alias{SCPattern-package}
\alias{SCPattern}
\docType{package}
\title{
SCPattern - single cell RNA-seq multiple condition data analysis 
}
\description{
SCPattern is a statistical method that classify genes into expression patterns in a
multiple condition single cell RNA-seq experiment.
}
\details{
	\tabular{ll}{
		Package: \tab SCPattern\cr
		Type: \tab Package\cr
		License: \tab Artistic-2.0\cr
	}
}
\author{
	Ning Leng

	Maintainer: Ning Leng <lengning1@gmail.com>
}
\references{
}
\keyword{ package }
