% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SCPTest.R
\name{SCPTest}
\alias{SCPTest}
\title{Perform directional or undirectional tests across two or more conditions}
\usage{
SCPTest <- function(Data, Conditions, sizeFactors,Directional = TRUE,
maxround=5, NumPat=3, Circular=FALSE,Dropout.remove=FALSE, Dropout.upper=0,
StateNames=c("EE","Down","Up"),UpdatePi=FALSE, Print=TRUE,SmallNum=1,  Seed=10, LOD=50)
}
\arguments{
\item{Data}{A data matrix contains expression values for each transcript
(gene or isoform level). In which rows should be transcripts
and columns should be samples. The matrix is expected to be unnormalized (raw data).}

\item{Conditions}{A factor indicates the condition which each sample belongs to.}

\item{sizeFactors}{The normalization factors. It should be a vector with lane
specific numbers (the length of the vector should be the same
as the number of samples, with the same order as the columns of Data).}

\item{Directional}{whether perform directional tests; default is TRUE.
If it is false, NumPat will be ignored.}

\item{Circular}{If Circular=TRUE, the function will also test last condition vs. first condition. Default is FALSE.}

\item{maxround}{Number of iterations. The default value is 5.}

\item{NumPat}{number of patterns to consider. Could be 2, 3 or 4. Default is 3.
When NumPat = 3, three states (no change (EE), down and up) are considered for each
transition. When NumPat = 2, two stages (down and up) are considered.
When NumPat = 4, four stages (no change, down, up and both directions) are considered.
NumPat will be ignored if Directional is FALSE. In this case the possible states are
no change and differentially expressed (DE).}

\item{UpdatePi}{default is FALSE}

\item{Print}{whether print the messages}

\item{SmallNum}{When calculating log2 expression, a small number was added to each of
the values. Default is 1.}

\item{Seed}{seed}

\item{LOD}{Genes with max value < LOD will be eliminated in the analysis.}

\item{method}{one of DirReg, VGAM and optim. It defines the method to use to estimate hyper-parameters}

\item{Droupout.remove,Dropout.upper}{For a given gene, whether cells with dropout will be
ignored in the test. Defaulte is FALSE. If it is set to TRUE, values that are less or equal to
Dropout.upper will be ignored during the test.}
}
\value{
Output: For each comparison, output posterior probabilities of being each state.
}
\description{
Perform directional or undirectional tests across two or more conditions
}
\examples{
mat <- matrix(exp(rnorm(100,10,1)),ncol=10)
Sizes <- MedianNorm(mat)
tmp <- SCPTest(mat, rep(c("c1","c2"),each=5), Sizes, LOD=0)
}
\author{
Ning Leng
}

