% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VioFun.R
\name{VioFun}
\alias{VioFun}
\title{Plot violinplot}
\usage{
VioFun(
  gene, Data, Conditions, Col, main_pre=NULL, Dropout.remove=FALSE,
  Dropout.upper=0, namecex=1, ylab="Expression", Log="", ...)
}
\arguments{
\item{gene}{gene name of interest (character)}

\item{Data}{input data, gene-by-sample matrix. Normalized expression is expected}

\item{Conditions}{a vector of characters that indicates which condition the cells belong to}

\item{Col}{colors}

\item{main_pre}{prefix for main}

\item{namecex}{cex for x axis names}

\item{ylab}{y axis label}

\item{Log}{whether plot in log scale}

\item{...}{additional arguments passed on to methods (e.g., ylim for vioplot).}

\item{Droupout.remove, Dropout.upper}{For a given gene, whether cells with dropout will be
ignored in the test. Defaulte is FALSE. If it is set to TRUE, values that are less or equal to
Dropout.upper will be ignored during the test.}
}
\value{
violin plot
}
\examples{
Data<- abs(matrix(rnorm(20,100,2), nrow=2))
rownames(Data) <- paste0("gene",1:2)
cond <- factor(rep(c("a","b"),each=5))
VioFun("gene1",Data, cond)
}
\author{
Ning Leng
}
